Fix compile.

+++ inc/extern.h
@@ -78,8 +78,8 @@ E void tb2sp(char *line);
 
 E char *strscpy(char *d, const char *s, size_t len);
 E void *smalloc(size_t size);
-E void *scalloc(long elsize, long els);
-E void *srealloc(void *oldptr, long newsize);
+E void *scalloc(size_t elsize, size_t els);
+E void *srealloc(void *oldptr, size_t newsize);
 E char *sstrdup(const char *s);
 E void strip(char *line);
 E char *make_hostmask(char *nick, char *user, char *host);
